--- scripts/config.js
@@ -1,11 +1,0 @@
-/*«-«-«-«-«-«-«-«-«-«-«-«-«-«-«-«-«-«-«-«-«-«-«-«-«-«-«-«-«-«-*/
-/*                          CONFIGS                           */
-/*-»-»-»-»-»-»-»-»-»-»-»-»-»-»-»-»-»-»-»-»-»-»-»-»-»-»-»-»-»-»*/
-
-const LATEST_SOLIDITY_VERSION = "0.8.24";
-const SRC_DIR = "src";
-
-module.exports = {
-  LATEST_SOLIDITY_VERSION,
-  SRC_DIR,
-};
